// desktop/src/generated/synkronus-client/models/AppBundleManifest.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * Synkronus API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.6
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AppBundleFile } from './AppBundleFile';
import {
    AppBundleFileFromJSON,
    AppBundleFileFromJSONTyped,
    AppBundleFileToJSON,
    AppBundleFileToJSONTyped,
} from './AppBundleFile';

/**
 * 
 * @export
 * @interface AppBundleManifest
 */
export interface AppBundleManifest {
    /**
     * 
     * @type {Array<AppBundleFile>}
     * @memberof AppBundleManifest
     */
    files: Array<AppBundleFile>;
    /**
     * 
     * @type {string}
     * @memberof AppBundleManifest
     */
    version: string;
    /**
     * 
     * @type {Date}
     * @memberof AppBundleManifest
     */
    generatedAt: Date;
    /**
     * 
     * @type {string}
     * @memberof AppBundleManifest
     */
    hash: string;
}

/**
 * Check if a given object implements the AppBundleManifest interface.
 */
export function instanceOfAppBundleManifest(value: object): value is AppBundleManifest {
    if (!('files' in value) || value['files'] === undefined) return false;
    if (!('version' in value) || value['version'] === undefined) return false;
    if (!('generatedAt' in value) || value['generatedAt'] === undefined) return false;
    if (!('hash' in value) || value['hash'] === undefined) return false;
    return true;
}

export function AppBundleManifestFromJSON(json: any): AppBundleManifest {
    return AppBundleManifestFromJSONTyped(json, false);
}

export function AppBundleManifestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AppBundleManifest {
    if (json == null) {
        return json;
    }
    return {
        
        'files': ((json['files'] as Array<any>).map(AppBundleFileFromJSON)),
        'version': json['version'],
        'generatedAt': (new Date(json['generatedAt'])),
        'hash': json['hash'],
    };
}

export function AppBundleManifestToJSON(json: any): AppBundleManifest {
    return AppBundleManifestToJSONTyped(json, false);
}

export function AppBundleManifestToJSONTyped(value?: AppBundleManifest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'files': ((value['files'] as Array<any>).map(AppBundleFileToJSON)),
        'version': value['version'],
        'generatedAt': value['generatedAt'].toISOString(),
        'hash': value['hash'],
    };
}